+++ lastversion/__about__.py
@@ -1,2 +1,2 @@
-__version__ = '2.4.6'
+__version__ = '2.4.7'
 __self__ = "dvershinin/lastversion"